Updated file.

+++ engines/ccgost/gost_lcl.h
@@ -189,9 +189,9 @@ int store_bignum(BIGNUM *bn, unsigned char *buf,int len);
 /* Read bignum, which can have few MSB all-zeros    from buffer*/ 
 BIGNUM *getbnfrombuf(const unsigned char *buf,size_t len);
 /* Pack GOST R 34.10 signature according to CryptoCom rules */
-int pack_sign_cc(DSA_SIG *s,int order,unsigned char *sig, unsigned int *siglen);
+int pack_sign_cc(DSA_SIG *s,int order,unsigned char *sig, size_t *siglen);
 /* Pack GOST R 34.10 signature according to CryptoPro rules */
-int pack_sign_cp(DSA_SIG *s,int order,unsigned char *sig, unsigned int *siglen); 
+int pack_sign_cp(DSA_SIG *s,int order,unsigned char *sig, size_t *siglen); 
 /* Unpack GOST R 34.10 signature according to CryptoCom rules */
 DSA_SIG *unpack_cc_signature(const unsigned char *sig,size_t siglen) ;
 /* Unpack GOST R 34.10 signature according to CryptoPro rules */
